python/requirements-direct.txt

alembic
Beaker
click
defusedxml
falcon
iso8601
jinja2
jsonschema
keystoneauth1
keystonemiddleware
libvirt-python
netaddr
oauthlib
oslo.config
oslo.policy
oslo.versionedobjects
Paste
PasteDeploy
psycopg2-binary
PTable
pyghmi
pylibyaml
pymongo
PyYAML
redfish
requests
SQLAlchemy<=1.3.20
ulid2
uWSGI